code/requirements.txt
groq>=0.9.0
google-genai>=1.0.0
rich>=13.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
tenacity>=8.2.0
